:recycle: Created folders for the redux actions
