+++ libs/authentication/lib/pages/signup_page.dart
@@ -4,7 +4,7 @@ import 'package:general/theme.dart';
 import 'package:general/widgets/blue_button_widget.dart';
 import 'package:general/widgets/dialog_helper.dart';
 import 'package:general/widgets/divider.dart';
-import 'package:redux_comp/actions/register_user_action.dart';
+import 'package:redux_comp/actions/user/register_user_action.dart';
 import 'package:redux_comp/redux_comp.dart';
 import '../widgets/button.dart';
 import '../widgets/circle_blur_widget.dart';
