+++ libs/authentication/lib/widgets/otp_pop_up.dart
@@ -3,7 +3,7 @@ import 'package:authentication/widgets/divider.dart';
 import 'package:authentication/widgets/textfield.dart';
 import 'package:flutter/material.dart';
 import 'package:general/widgets/blue_button_widget.dart';
-import 'package:redux_comp/actions/verify_user_action.dart';
+import 'package:redux_comp/actions/user/verify_user_action.dart';
 import 'package:redux_comp/app_state.dart';
 
 import '../pages/login_page.dart';
