+++ libs/consumer/lib/pages/bid_details_page.dart
@@ -4,8 +4,8 @@ import 'package:general/general.dart';
 import 'package:general/widgets/appbar.dart';
 import 'package:general/widgets/floating_button.dart';
 import 'package:general/widgets/navbar.dart';
-import 'package:redux_comp/actions/accept_bid_action.dart';
-import 'package:redux_comp/actions/shortlist_bid_action.dart';
+import 'package:redux_comp/actions/bids/accept_bid_action.dart';
+import 'package:redux_comp/actions/bids/shortlist_bid_action.dart';
 import 'package:redux_comp/app_state.dart';
 import 'package:general/widgets/shortlist_accept_button.dart';
 import 'package:redux_comp/models/bid_model.dart';
