+++ libs/consumer/lib/pages/create_new_advert_page.dart
@@ -5,7 +5,7 @@ import 'package:general/widgets/appbar.dart';
 import 'package:general/widgets/button.dart';
 import 'package:general/widgets/navbar.dart';
 import 'package:general/widgets/textfield.dart';
-import 'package:redux_comp/actions/create_advert_action.dart';
+import 'package:redux_comp/actions/adverts/create_advert_action.dart';
 import 'package:redux_comp/redux_comp.dart';
 import 'package:general/widgets/floating_button.dart';
