+++ libs/redux_comp/lib/actions/adverts/close_advert_action.dart
@@ -1,6 +1,6 @@
 import 'package:amplify_api/amplify_api.dart';
 
-import '../app_state.dart';
+import '../../app_state.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:async_redux/async_redux.dart';
