+++ libs/redux_comp/lib/actions/adverts/create_advert_action.dart
@@ -3,9 +3,9 @@ import 'dart:convert';
 import 'package:amplify_api/amplify_api.dart';
 import 'package:redux_comp/models/advert_model.dart';
 import 'package:uuid/uuid.dart';
-import '../app_state.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:async_redux/async_redux.dart';
+import '../../app_state.dart';
 
 // creates an advert
 // requires the customerdId and a title the rest is optional
