+++ libs/redux_comp/lib/actions/adverts/edit_advert_action.dart
@@ -1,7 +1,7 @@
 import 'package:amplify_api/amplify_api.dart';
 import 'package:redux_comp/models/advert_model.dart';
 
-import '../app_state.dart';
+import '../../app_state.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:async_redux/async_redux.dart';
