+++ libs/redux_comp/lib/actions/adverts/set_active_ad_action.dart
@@ -1,5 +1,5 @@
 import 'package:redux_comp/models/advert_model.dart';
-import '../app_state.dart';
+import '../../app_state.dart';
 import 'package:async_redux/async_redux.dart';
 
 class SetActiveAdAction extends ReduxAction<AppState> {
