+++ libs/redux_comp/lib/actions/adverts/view_adverts_action.dart
@@ -1,18 +1,21 @@
+import 'dart:async';
 import 'dart:convert';
 
 import 'package:amplify_api/amplify_api.dart';
-
-import '../app_state.dart';
+import 'package:redux_comp/models/advert_model.dart';
+import '../../app_state.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:async_redux/async_redux.dart';
 
-import '../models/advert_model.dart';
+class ViewAdvertsAction extends ReduxAction<AppState> {
+  final String consId; // consumer id whos adverts you wish to retrieve
 
-class ViewJobsAction extends ReduxAction<AppState> {
-	@override
-	Future<AppState?> reduce() async {
+  ViewAdvertsAction(this.consId);
+
+  @override
+  Future<AppState?> reduce() async {
     String graphQLDocument = '''query {
-      viewJobs(locations: "Pretoria", type: "Plumbing") {
+      viewAdverts(user_id: "$consId") {
         date_created
         date_closed
         description
@@ -30,7 +33,7 @@ class ViewJobsAction extends ReduxAction<AppState> {
       final response = await Amplify.API.query(request: request).response;
 
       List<AdvertModel> adverts = [];
-      dynamic data = jsonDecode(response.data)['viewJobs'];
+      dynamic data = jsonDecode(response.data)['viewAdverts'];
       data.forEach((el) => adverts.add(AdvertModel.fromJson(el)));
 
       return state.replace(
