+++ libs/redux_comp/lib/actions/bids/accept_bid_action.dart
@@ -1,10 +1,10 @@
 import 'package:amplify_api/amplify_api.dart';
 import 'package:redux_comp/models/bid_model.dart';
-import '../app_state.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:async_redux/async_redux.dart';
 
-import '../models/bid_model.dart';
+import '../../models/bid_model.dart';
+import '../../app_state.dart';
 
 class AcceptBidAction extends ReduxAction<AppState> {
   AcceptBidAction();
