+++ libs/redux_comp/lib/actions/bids/delete_bid_action.dart
@@ -1,7 +1,7 @@
 import 'package:amplify_api/amplify_api.dart';
 import 'package:flutter/cupertino.dart';
 
-import '../app_state.dart';
+import '../../app_state.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:async_redux/async_redux.dart';
