+++ libs/redux_comp/lib/actions/bids/set_active_bid_action.dart
@@ -1,5 +1,5 @@
 import 'package:redux_comp/models/bid_model.dart';
-import '../app_state.dart';
+import '../../app_state.dart';
 import 'package:async_redux/async_redux.dart';
 
 class SetActiveBidAction extends ReduxAction<AppState> {
