+++ libs/redux_comp/lib/actions/bids/toggle_view_bids_action.dart
@@ -1,5 +1,5 @@
 import 'package:redux_comp/models/bid_model.dart';
-import '../app_state.dart';
+import '../../app_state.dart';
 import 'package:async_redux/async_redux.dart';
 
 class ToggleViewBidsAction extends ReduxAction<AppState> {
