+++ libs/redux_comp/lib/actions/geolocation/get_place_action.dart
@@ -1,7 +1,7 @@
 import 'package:redux_comp/models/geolocation/coordinates_model.dart';
 import 'package:redux_comp/models/geolocation/place_model.dart';
 
-import '../app_state.dart';
+import '../../app_state.dart';
 import 'package:async_redux/async_redux.dart';
 
 class GetPlaceAction extends ReduxAction<AppState> {
