+++ libs/redux_comp/lib/actions/geolocation/get_suggestions_action.dart
@@ -1,6 +1,6 @@
 import 'package:redux_comp/models/geolocation/suggestion_model.dart';
 
-import '../app_state.dart';
+import '../../app_state.dart';
 import 'package:async_redux/async_redux.dart';
 import 'package:geolocation/place_api_service.dart';
