+++ libs/redux_comp/lib/actions/user/add_user_to_group_action.dart
@@ -1,10 +1,10 @@
 import 'package:amplify_api/amplify_api.dart';
 import 'package:flutter/material.dart';
-import '../app_state.dart';
+import '../../app_state.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:async_redux/async_redux.dart';
 
-import '../models/error_type_model.dart';
+import '../../models/error_type_model.dart';
 
 class AddUserToGroupAction extends ReduxAction<AppState> {
   @override
