+++ libs/redux_comp/lib/actions/user/create_user_action.dart
@@ -2,10 +2,10 @@ import 'dart:convert';
 
 import 'package:amplify_api/amplify_api.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:redux_comp/actions/login_action.dart';
+import 'package:redux_comp/actions/user/login_action.dart';
 import 'package:redux_comp/models/error_type_model.dart';
 
-import '../app_state.dart';
+import '../../app_state.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:async_redux/async_redux.dart';
