+++ libs/redux_comp/lib/actions/user/login_action.dart
@@ -3,9 +3,9 @@ import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:async_redux/async_redux.dart';
 import 'package:flutter/foundation.dart';
-import 'package:redux_comp/actions/get_user_action.dart';
-import 'package:redux_comp/actions/view_adverts_action.dart';
-import 'package:redux_comp/actions/view_jobs_action.dart';
+import 'package:redux_comp/actions/adverts/view_adverts_action.dart';
+import 'package:redux_comp/actions/adverts/view_jobs_action.dart';
+import 'package:redux_comp/actions/user/get_user_action.dart';
 import 'package:redux_comp/app_state.dart';
 import 'package:redux_comp/models/error_type_model.dart';
 import 'package:jwt_decode/jwt_decode.dart';
