+++ libs/redux_comp/lib/actions/user/verify_user_action.dart
@@ -1,10 +1,10 @@
 import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
 import 'package:flutter/foundation.dart';
-import 'package:redux_comp/actions/add_user_to_group_action.dart';
-import 'package:redux_comp/actions/create_user_action.dart';
-import '../app_state.dart';
+import 'package:redux_comp/actions/user/add_user_to_group_action.dart';
+import 'package:redux_comp/actions/user/create_user_action.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:async_redux/async_redux.dart';
+import '../../app_state.dart';
 
 class VerifyUserAction extends ReduxAction<AppState> {
   final String confirmationCode;
