+++ libs/redux_comp/lib/redux_comp.dart
@@ -2,4 +2,3 @@ library redux;
 // file should just export redux pattern subsytems
 
 export './app_state.dart';
-export './actions/login_action.dart';
