+++ libs/tradesman/lib/pages/view_bids_page.dart
@@ -10,10 +10,10 @@ import 'package:general/widgets/button.dart';
 import 'package:general/widgets/navbar.dart';
 import 'package:flutter/material.dart';
 import 'package:general/widgets/job_card.dart';
+import 'package:redux_comp/actions/bids/toggle_view_bids_action.dart';
 import 'package:redux_comp/app_state.dart';
 import 'package:redux_comp/models/advert_model.dart';
 import 'package:redux_comp/models/bid_model.dart';
-import 'package:redux_comp/actions/toggle_view_bids_action.dart';
 import 'package:general/widgets/floating_button.dart';
 
 class TradesmanViewBidsPage extends StatelessWidget {
