+++ libs/tradesman/lib/widgets/place_bid_popup.dart
@@ -1,7 +1,7 @@
 import 'package:async_redux/async_redux.dart';
 import 'package:authentication/widgets/divider.dart';
 import 'package:flutter/material.dart';
-import 'package:redux_comp/actions/place_bid_action.dart';
+import 'package:redux_comp/actions/bids/place_bid_action.dart';
 import 'package:redux_comp/app_state.dart';
 import 'package:redux_comp/models/advert_model.dart';
